Renaming vaguely named file.
Also adding an UPDATE_DRAFTS action.
